Major improvements to Create Tournament in Tournament.java & Improved output in Countries.java
Creating a tournament will now be saved into the games table in Olympics database.

Countries.java has a better country list output in which that it will only output countries that actually has a list of Participants in the participants. This prevents selecting a country without any Participants. Will add an user option later on so that the user can decide which output they would like.

public boolean addNewTournament function was created in Tournament.java

Foreign table countryInGames was created as well to work with tournament management.

Create tournament is still incomplete due to small issues with inserting values in countryInGames